refactor: moved creation of lidars_and_cameras_sequence into conversion, and load at upload
