+++ config/convert_annotated_t4_to_kognic_sample.yaml
@@ -3,6 +3,7 @@ conversion:
   input_base: ./data/annotated_t4_format
   output_base: ./data/kognic_format
   workers_number: 12
+  include_imu_data: false
   camera_sensors:
     - channel: CAM_FRONT
     - channel: CAM_FRONT_RIGHT
