+++ perception_dataset/kognic/t4_to_kognic_converter.py
@@ -6,12 +6,32 @@ import json
 from pathlib import Path
 import shutil
 import time
-from typing import Dict, List, Set, Tuple
+from typing import Dict, Generator, List, Optional, Set, Tuple
+import uuid
 
+import kognic.io.model as KognicModel
+from kognic.io.model.ego.imu_data import IMUData
+from kognic.io.model.scene.lidars_and_cameras_sequence.frame import (
+    Frame as LidarsAndCamerasSequenceFrame,
+)
+from kognic.io.model.scene.metadata.metadata import FrameMetaData, MetaData
+from kognic.io.model.scene.resources.image import ImageMetadata
+import numpy as np
 from t4_devkit import Tier4
 
 from perception_dataset.abstract_converter import AbstractConverter
-from perception_dataset.constants import LIDAR_CONCAT_CHANNEL
+from perception_dataset.constants import (
+    IMU_EXTRAPOLATE_S,
+    IMU_TARGET_HZ,
+    LIDAR_CONCAT_CHANNEL,
+    PREFERRED_CAMERA_SENSORS,
+    PREFERRED_LIDAR_SENSORS,
+)
+from perception_dataset.kognic.sequence_artifact import (
+    PENDING_CALIBRATION_ID,
+    SEQUENCE_ARTIFACT_FILENAME,
+    save_sequence_artifact,
+)
 from perception_dataset.kognic.utils import (
     extract_calibration,
     extract_ego_poses,
@@ -39,6 +59,76 @@ _REPORT_FIELDS = (
 )
 
 
+def _sort_key(path: Path) -> Tuple[int, str]:
+    """Sort timestamp-named sensor files numerically, then other names lexically."""
+    try:
+        return (0, f"{int(path.stem):030d}")
+    except ValueError:
+        return (1, path.stem)
+
+
+def _sensor_sort_key(sensor_name: str, preferred_order: List[str]) -> Tuple[int, str]:
+    """Prioritize known sensors while keeping deterministic fallback order."""
+    if sensor_name in preferred_order:
+        return (preferred_order.index(sensor_name), sensor_name)
+    return (len(preferred_order), sensor_name)
+
+
+def _validate_sensor_file_counts(
+    sequence_path: Path,
+    sensor_files: Dict[str, List[Path]],
+    anchor_sensor: str,
+    expected_count: int,
+) -> None:
+    """Require every sensor to contribute exactly one file per frame."""
+    mismatched = {
+        name: len(files) for name, files in sensor_files.items() if len(files) != expected_count
+    }
+    if mismatched:
+        raise ValueError(
+            f"Sensor file counts disagree in {sequence_path}: anchor {anchor_sensor} has "
+            f"{expected_count} files but {mismatched} differ. Frames are paired with files by "
+            "position, so re-convert the sequence instead of uploading a partial staging directory."
+        )
+
+
+def _validate_sensor_timestamps(
+    sequence_path: Path,
+    sensor_files: Dict[str, List[Path]],
+    anchor_files: List[Path],
+) -> None:
+    """Require each sensor's timestamps to track the anchor sensor's frames."""
+    try:
+        anchor_ts = [int(path.stem) for path in anchor_files]
+    except ValueError:
+        return
+
+    intervals = sorted(b - a for a, b in zip(anchor_ts, anchor_ts[1:]))
+    if not intervals:
+        return
+
+    tolerance_ns = intervals[len(intervals) // 2] * 0.9
+    if tolerance_ns <= 0:
+        raise ValueError(
+            f"Anchor timestamps in {sequence_path} are not strictly increasing; "
+            "frame ordering cannot be trusted."
+        )
+
+    for name, files in sensor_files.items():
+        for frame_idx, (path, expected_ns) in enumerate(zip(files, anchor_ts)):
+            try:
+                actual_ns = int(path.stem)
+            except ValueError:
+                break
+            if abs(actual_ns - expected_ns) > tolerance_ns:
+                raise ValueError(
+                    f"Sensor {name} in {sequence_path} is misaligned at frame {frame_idx}: "
+                    f"file {path.name} is {abs(actual_ns - expected_ns)}ns from the anchor "
+                    f"timestamp {expected_ns} (tolerance {tolerance_ns:.0f}ns). The staging "
+                    "directory likely mixes files from different conversion runs."
+                )
+
+
 class T4ToKognicConverter(AbstractConverter[None]):
     """Convert T4 data (annotated or non-annotated) to the Kognic IO staging layout.
 
@@ -51,16 +141,16 @@ class T4ToKognicConverter(AbstractConverter[None]):
             calibration.json
             ego_poses.json
             keyframes.json
+            lidars_and_cameras_sequence.json
             cameras/<camera_name>/<timestamp_ns>.jpg
             lidar/<lidar_name>/<timestamp_ns>.csv
 
     ``keyframes.json`` holds the staging frame indices of the keyframes; the
     uploader requires this file and marks exactly those frames ``annotate=True``.
-    For annotated datasets (``annotated=True``) the
-    keyframes are the frames whose sample carries at least one
-    ``sample_annotation``. Non-annotated datasets have no annotations, so their
-    keyframes are instead selected by sample index at ``annotation_hz``,
-    matching the non-annotated T4 -> Deepen converter.
+    For annotated datasets (``annotated=True``), source T4 keyframes remain
+    annotatable even when they contain no objects. Non-annotated datasets have
+    no source keyframes, so their keyframes are selected by sample index at
+    ``annotation_hz``, matching the non-annotated T4 -> Deepen converter.
     """
 
     def __init__(
@@ -71,6 +161,7 @@ class T4ToKognicConverter(AbstractConverter[None]):
         workers_number: int = 32,
         annotated: bool = True,
         annotation_hz: int = 10,
+        include_imu_data: bool = True,
         generate_tsv_report: bool = False,
     ):
         """Initialize the converter.
@@ -83,6 +174,8 @@ class T4ToKognicConverter(AbstractConverter[None]):
             annotated (bool): Whether the source carries T4 annotations.
             annotation_hz (int): Keyframe frequency for non-annotated data, in
                 ``1..10``.
+            include_imu_data (bool): Whether to include interpolated IMU data
+                in the validated sequence artifact.
             generate_tsv_report (bool): Write ``conversion_report.tsv`` in
                 ``output_base`` with scene outcomes and missing sensor frames.
 
@@ -94,6 +187,7 @@ class T4ToKognicConverter(AbstractConverter[None]):
         self._workers_number = workers_number
         self._annotated = annotated
         self._annotation_hz = validate_annotation_hz(annotation_hz)
+        self._include_imu_data = include_imu_data
         self._generate_tsv_report = generate_tsv_report
         self._report_rows: List[Dict[str, str]] = []
         # Cache one blank black image per camera, sized to that camera's frames,
@@ -164,13 +258,17 @@ class T4ToKognicConverter(AbstractConverter[None]):
         out_dir = Path(output_dir)
         out_dir.mkdir(parents=True, exist_ok=True)
 
-        # The uploader pairs sensor files with frames by sorted position, so any
+        # Sequence construction pairs sensor files by sorted position, so any
         # file left over from a previous generation would shift that pairing.
         for stale_dir in (out_dir / "cameras", out_dir / "lidar"):
             shutil.rmtree(stale_dir, ignore_errors=True)
+        (out_dir / SEQUENCE_ARTIFACT_FILENAME).unlink(missing_ok=True)
 
         self._build_lookup_maps(seq_path)
-        self._has_lidar_concat_info = (seq_path / "data" / "LIDAR_CONCAT_INFO").is_dir()
+        self._has_lidar_concat_info = any(
+            sample_data.info_filename
+            for sample_data in self._sample_data_by_channel.get(LIDAR_CONCAT_CHANNEL, [])
+        )
         self._lidar_channels = self._discover_lidar_channels()
         self._frame_records = self._build_frame_records()
         self._record_missing_sensor_frames(seq_path)
@@ -219,6 +317,8 @@ class T4ToKognicConverter(AbstractConverter[None]):
                 frame_records=self._frame_records,
                 channel_to_token=self._channel_to_token,
             )
+
+        self._write_sequence_artifact(out_dir, ego_poses)
 
     # ------------------------------------------------------------------
     # Conversion report
@@ -417,12 +517,6 @@ class T4ToKognicConverter(AbstractConverter[None]):
         samples = t4.get_table("sample")
         self._samples = sorted(samples, key=lambda s: s.timestamp)
 
-        self._annotated_sample_tokens: Set[str] = set()
-        if self._annotated:
-            self._annotated_sample_tokens = {
-                annotation.sample_token for annotation in t4.get_table("sample_annotation")
-            }
-
         self._sample_data_by_channel: Dict[str, list] = {}
         self._sample_data_by_channel_and_frame_id: Dict[str, Dict[str, object]] = {}
         for sd in t4.get_table("sample_data"):
@@ -483,8 +577,8 @@ class T4ToKognicConverter(AbstractConverter[None]):
         ``frame_count`` lets it detect a stale file after the staging data
         changed.
 
-        Annotated datasets: a frame is a keyframe when its sample carries at
-        least one ``sample_annotation``.
+        Annotated datasets: source T4 keyframes remain annotatable, including
+        keyframes that contain no objects.
 
         Non-annotated datasets: there are no annotations to key off, so
         keyframes are selected by sample index at ``annotation_hz``, with the
@@ -501,8 +595,7 @@ class T4ToKognicConverter(AbstractConverter[None]):
             keyframe_indices = [
                 idx
                 for idx, frame_record in enumerate(self._frame_records)
-                if getattr(frame_record.get(self._anchor_channel), "sample_token", None)
-                in self._annotated_sample_tokens
+                if getattr(frame_record.get(self._anchor_channel), "is_key_frame", False)
             ]
         else:
             step = int(MAX_ANNOTATION_HZ / self._annotation_hz)
@@ -525,6 +618,261 @@ class T4ToKognicConverter(AbstractConverter[None]):
         logger.info(
             f"keyframes.json: {len(keyframe_indices)} keyframes over "
             f"{len(self._frame_records)} frames"
+        )
+
+    @staticmethod
+    def _collect_sensor_files(
+        sequence_path: Path, root_name: str, suffix: str
+    ) -> Dict[str, List[Path]]:
+        """Collect converted files grouped by sensor channel."""
+        root = sequence_path / root_name
+        if not root.exists():
+            return {}
+
+        sensor_files = {}
+        for sensor_dir in sorted(path for path in root.iterdir() if path.is_dir()):
+            files = sorted(sensor_dir.glob(f"*{suffix}"), key=_sort_key)
+            if files:
+                sensor_files[sensor_dir.name] = files
+        return sensor_files
+
+    def _iterate_frames(
+        self, sequence_path: Path
+    ) -> Generator[Tuple[str, int, Dict[str, Path]], None, None]:
+        """Yield synchronized converted frames in anchor-sensor order."""
+        lidar_files = self._collect_sensor_files(sequence_path, "lidar", ".csv")
+        camera_files = self._collect_sensor_files(sequence_path, "cameras", ".jpg")
+        if not lidar_files and not camera_files:
+            raise FileNotFoundError(f"No lidar CSVs or camera JPGs found in {sequence_path}")
+
+        if lidar_files:
+            anchor_sensors = sorted(
+                lidar_files,
+                key=lambda name: _sensor_sort_key(name, PREFERRED_LIDAR_SENSORS),
+            )
+            anchor_files = lidar_files[anchor_sensors[0]]
+        else:
+            anchor_sensors = sorted(
+                camera_files,
+                key=lambda name: _sensor_sort_key(name, PREFERRED_CAMERA_SENSORS),
+            )
+            anchor_files = camera_files[anchor_sensors[0]]
+
+        combined_files = {**lidar_files, **camera_files}
+        all_files = {
+            name: combined_files[name]
+            for name in [
+                anchor_sensors[0],
+                *sorted(set(combined_files) - {anchor_sensors[0]}),
+            ]
+        }
+        _validate_sensor_file_counts(
+            sequence_path,
+            all_files,
+            anchor_sensors[0],
+            len(anchor_files),
+        )
+        _validate_sensor_timestamps(sequence_path, all_files, anchor_files)
+
+        for frame_idx, anchor_file in enumerate(anchor_files):
+            yield (
+                str(frame_idx),
+                int(anchor_file.stem),
+                {name: files[frame_idx] for name, files in all_files.items()},
+            )
+
+    @staticmethod
+    def _load_keyframe_indices(sequence_path: Path, frame_count: int) -> List[int]:
+        """Load and validate converted T4 keyframe positions."""
+        keyframes_path = sequence_path / "keyframes.json"
+        if not keyframes_path.exists():
+            raise FileNotFoundError(f"Required keyframe metadata is missing: {keyframes_path}")
+
+        with open(keyframes_path) as f:
+            data = json.load(f)
+        if data.get("frame_count") != frame_count:
+            raise ValueError(
+                f"{keyframes_path} was generated for {data.get('frame_count')} frames but "
+                f"the current staging data has {frame_count}."
+            )
+
+        keyframe_indices = data.get("keyframe_indices")
+        if not isinstance(keyframe_indices, list) or any(
+            not isinstance(index, int) or isinstance(index, bool) or not 0 <= index < frame_count
+            for index in keyframe_indices
+        ):
+            raise ValueError(
+                f"{keyframes_path} contains invalid keyframe_indices for {frame_count} frames."
+            )
+        return keyframe_indices
+
+    def _build_sequence_frames(
+        self,
+        sequence_path: Path,
+        ego_poses: Dict[str, KognicModel.EgoVehiclePose],
+    ) -> List[LidarsAndCamerasSequenceFrame]:
+        """Build validated Kognic frames from converted sensor files."""
+        frames = []
+        frame_records = list(self._iterate_frames(sequence_path))
+        annotate_indices = set(
+            self._load_keyframe_indices(sequence_path, len(frame_records))
+        )
+        reference_timestamp: Optional[int] = None
+
+        for frame_idx, (frame_id, timestamp_ns, sensor_files) in enumerate(frame_records):
+            if reference_timestamp is None:
+                reference_timestamp = timestamp_ns
+            frames.append(
+                LidarsAndCamerasSequenceFrame(
+                    frame_id=frame_id,
+                    relative_timestamp=int((timestamp_ns - reference_timestamp) / 1e6),
+                    unix_timestamp=timestamp_ns,
+                    ego_vehicle_pose=ego_poses.get(frame_id),
+                    point_clouds=[
+                        KognicModel.PointCloud(sensor_name=name, filename=str(path))
+                        for name, path in sensor_files.items()
+                        if path.suffix == ".csv"
+                    ],
+                    images=[
+                        KognicModel.Image(
+                            sensor_name=name,
+                            filename=str(path),
+                            metadata=ImageMetadata(
+                                shutter_time_start_ns=int(path.stem),
+                                shutter_time_end_ns=int(path.stem),
+                            ),
+                        )
+                        for name, path in sensor_files.items()
+                        if path.suffix == ".jpg"
+                    ],
+                    metadata=FrameMetaData(annotate=frame_idx in annotate_indices),
+                )
+            )
+        return frames
+
+    def _build_imu_data(
+        self,
+        sequence_path: Path,
+        ego_poses: Dict[str, KognicModel.EgoVehiclePose],
+    ) -> List[IMUData]:
+        """Interpolate converted ego poses into dense Kognic IMU data."""
+        if not self._include_imu_data or not ego_poses:
+            return []
+
+        try:
+            from scipy.interpolate import interp1d
+            from scipy.spatial.transform import Rotation, Slerp
+        except ModuleNotFoundError:
+            logger.warning("scipy is not installed; skipping optional IMU data generation")
+            return []
+
+        sparse = [
+            (timestamp_ns, ego_poses[frame_id])
+            for frame_id, timestamp_ns, _ in self._iterate_frames(sequence_path)
+            if frame_id in ego_poses
+        ]
+        if len(sparse) < 2:
+            return []
+
+        sparse.sort(key=lambda item: item[0])
+        ts_sparse = np.array([item[0] for item in sparse], dtype=np.float64)
+        pos_sparse = np.array(
+            [
+                [item[1].position.x, item[1].position.y, item[1].position.z]
+                for item in sparse
+            ],
+            dtype=np.float64,
+        )
+        rot_sparse = Rotation.concatenate(
+            [
+                Rotation.from_quat(
+                    [
+                        item[1].rotation.x,
+                        item[1].rotation.y,
+                        item[1].rotation.z,
+                        item[1].rotation.w,
+                    ]
+                )
+                for item in sparse
+            ]
+        )
+
+        dense_dt_ns = int(1e9 / IMU_TARGET_HZ)
+        extrap_ns = IMU_EXTRAPOLATE_S * 1e9
+        ts_dense = np.arange(
+            ts_sparse[0] - extrap_ns,
+            ts_sparse[-1] + extrap_ns + dense_dt_ns,
+            dense_dt_ns,
+        )
+        pos_dense = interp1d(
+            ts_sparse,
+            pos_sparse,
+            axis=0,
+            fill_value="extrapolate",
+            bounds_error=False,
+        )(ts_dense)
+        slerp = Slerp(ts_sparse, rot_sparse)
+        rot_dense_list = []
+        for timestamp in ts_dense:
+            if timestamp <= ts_sparse[0]:
+                delta = ts_sparse[1] - ts_sparse[0]
+                difference = rot_sparse[0].inv() * rot_sparse[1]
+                rot_dense_list.append(
+                    rot_sparse[0]
+                    * (difference.inv() ** ((ts_sparse[0] - timestamp) / delta))
+                )
+            elif timestamp >= ts_sparse[-1]:
+                delta = ts_sparse[-1] - ts_sparse[-2]
+                difference = rot_sparse[-2].inv() * rot_sparse[-1]
+                rot_dense_list.append(
+                    rot_sparse[-1]
+                    * (difference ** ((timestamp - ts_sparse[-1]) / delta))
+                )
+            else:
+                rot_dense_list.append(slerp(timestamp))
+
+        quaternions = Rotation.concatenate(rot_dense_list).as_quat()
+        return [
+            IMUData(
+                timestamp=float(ts_dense[index]),
+                position=KognicModel.Position(
+                    x=float(pos_dense[index, 0]),
+                    y=float(pos_dense[index, 1]),
+                    z=float(pos_dense[index, 2]),
+                ),
+                rotation_quaternion=KognicModel.RotationQuaternion(
+                    w=float(quaternions[index, 3]),
+                    x=float(quaternions[index, 0]),
+                    y=float(quaternions[index, 1]),
+                    z=float(quaternions[index, 2]),
+                ),
+            )
+            for index in range(len(ts_dense))
+        ]
+
+    def _write_sequence_artifact(
+        self,
+        sequence_path: Path,
+        ego_poses: Dict[str, KognicModel.EgoVehiclePose],
+    ) -> None:
+        """Build, validate, and persist the complete Kognic sequence."""
+        frames = self._build_sequence_frames(sequence_path, ego_poses)
+        imu_data = self._build_imu_data(sequence_path, ego_poses)
+        sequence = KognicModel.LidarsAndCamerasSequence(
+            external_id=sequence_path.name,
+            frames=frames,
+            calibration_id=PENDING_CALIBRATION_ID,
+            imu_data=imu_data,
+            metadata=MetaData(
+                source_filename=sequence_path.name,
+                dataset_id=sequence_path.name,
+                inner_uuid=str(uuid.uuid5(uuid.NAMESPACE_URL, sequence_path.name)),
+            ),
+        )
+        save_sequence_artifact(sequence_path, sequence)
+        logger.info(
+            f"{SEQUENCE_ARTIFACT_FILENAME}: {len(frames)} frames, "
+            f"{len(imu_data)} IMU samples"
         )
 
     def _build_frame_records(self) -> List[Dict[str, object]]:
@@ -633,7 +981,7 @@ class T4ToKognicConverter(AbstractConverter[None]):
                 # correct frame position at upload time.
                 timestamp_ns = self._frame_timestamp_ns(frame_record)
 
-            # Frames must map 1:1 onto files. The uploader rebuilds frames by
+            # Frames must map 1:1 onto files. Sequence construction builds frames by
             # counting files per sensor directory and pairing them by index, so a
             # dropped file shifts every later frame onto the wrong image and
             # leaves the last frame without one, failing Kognic scene validation
